upload ReactYouTubeProfile
ReactYouTubeProfile is a React-based project that simulates a YouTube profile page. It features components for displaying and updating a user's banner and profile images, with the ability to save these images in local storage. The project also includes a sidebar for navigation and a profile section with user details and subscription options.